requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.6

# Slack
slack-bolt==1.18.1
slack-sdk==3.26.2

# Google Ads
google-ads==23.1.0

# Google OAuth & API
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Google Gemini AI
google-generativeai==0.3.2

# Security
cryptography==42.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0,<5.0  # passlib 1.7.4 is not compatible with bcrypt 5.x

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
